Remove string interning by default
It is still available via the `StringInterningFormatter`, which can be turned on by default by aggregating a custom resolver,
or by specifying this on a `[Key]` or `[DataMember]` member:

```cs
[MessagePackFormatter(typeof(StringInterningFormatter))]
```
